Create the salk_toolkit.io package (pure move)
Split salk_toolkit/io.py (2397 lines) into a salk_toolkit/io/ package by
relocation only. Every one of the 52 top-level functions/classes is
byte-identical to main (verified by an AST body-diff); the package
__init__ re-exports the exact public surface plus the private helpers that
tests and salk_internal_package import, so no test and no caller changes.

Module layout: core (shared series helpers), readers (file tracking/map),
meta (DataMeta introspection + category fixing), parquet (embedded-meta
read/write), create_blocks (topk/maxdiff), sources (_load_data_files +
_reconcile_categories), pipeline (_process_annotated_data), datasets
(read_annotated_data / read_and_process_data / infer_meta / merges).

The only deviations from verbatim are the mechanical necessities of
crossing a module boundary:
- stk_file_map / stk_loaded_files_set are accessed as readers.<attr> in
  sources and pipeline; a 'from readers import stk_file_map' would capture
  a stale binding and silently break set_file_map's rebind.
- a deferred 'from salk_toolkit.io.datasets import read_annotated_data'
  inside _load_data_files breaks the single sources<->datasets import
  cycle (removed properly in the follow-up via _load_dataset).
- per-module import headers.

Part of specs/io-pipeline-refactor-design.md, phase (a).
