+++ salk_toolkit/io/readers.py
@@ -1,0 +1,48 @@
+"""File tracking and path remapping for reproducible packaging."""
+
+
+# This is here so we can easily track which files would be needed for a model
+# so we can package them together if needed
+
+#  a global list of files that have been loaded
+stk_loaded_files_set = set()
+
+
+def get_loaded_files() -> list[str]:
+    """Get list of all files that have been loaded during this session.
+
+    Returns:
+        List of file paths that have been loaded.
+    """
+    global stk_loaded_files_set
+    return list(stk_loaded_files_set)
+
+
+def reset_file_tracking() -> None:
+    """Clear the set of tracked loaded files."""
+    global stk_loaded_files_set
+    stk_loaded_files_set.clear()
+
+
+# a global map that allows remapping file paths/names to different paths
+stk_file_map = {}
+
+
+def get_file_map() -> dict[str, str]:
+    """Get the current file path mapping dictionary.
+
+    Returns:
+        Copy of the file map dictionary.
+    """
+    global stk_file_map
+    return stk_file_map.copy()
+
+
+def set_file_map(file_map: dict[str, str]) -> None:
+    """Set the file path mapping dictionary.
+
+    Args:
+        file_map: Dictionary mapping original paths to new paths.
+    """
+    global stk_file_map
+    stk_file_map = file_map.copy()
